Add PokeMMO
While this is functional, there are a couple of issues with it still:

- [ ] No permission to use icons from upstream
- [ ] File verification fails on start
- [ ] The StartupWMClass may be wrong
- [ ] Portals aren't used, so supplying ROMs is hard
- [ ] x-checker-data is untested
- [ ] Translations are missing
- [ ] No screenshots or other niceties

Either way, I need to contact the upstream and see if they're
interested in publishing this. Mostly for my own use for now, but I
figured I'd push this up so that other interested people can gather
round and see if we can convince upstream.